Create new folder for html widgets and first widget
This commit adds in a folder for the future storage of all widgets for
more organized file management.

It also adds the first html widget for this branch: ntee1_table.
This is an interactive table to explore the original IRS Activity codes
and is pending further design input from the NCCS team. The table will
make parsing old codes easier and will also keep users on the new nccs
site instead of continuing to reroute them to the old site.