Add nccs-site-diagrams.pptx to public images asset folder
I combined all of the diagrams that are used in NCCS site documentation into a single power point file that can live in the "public/img" web assets folder. If the images need to be updated, change them in the PPTX file, save changes, then right-click and save-as a PNG image in the appropriate folder.